closed #381

+++ database/migrations/2019_10_31_084035_create_loans_table.php
@@ -16,7 +16,7 @@ class CreateLoansTable extends Migration
         Schema::create('loans', function (Blueprint $table) {
             $table->bigIncrements('id');// id unico
             $table->string('code')->nullable(); // para el correlativo
-            $table->integer('disbursable_id');// id affiliado, id espouse
+            $table->unsignedBigInteger('disbursable_id');// id affiliado, id espouse
             $table->enum('disbursable_type',['affiliates', 'spouses']); // a quien se hara del desembolso//afiliado, conyugue
             $table->unsignedBigInteger('procedure_modality_id'); // id modalidad
             $table->foreign('procedure_modality_id')->references('id')->on('procedure_modalities');
